Added operations for the spectrum class
Added common arithmetic operations as well as numpy functions to be used
in the SPECTRUM class. To do this we defined the __array_ufunc__ method
as recommended here

https://numpy.org/doc/stable/reference/arrays.classes.html

as well as changed the kind to numpy.lib.mixins.NDArrayOperatorsMixin to
allow for the normal operations to be handled by numpy.

An SPECTRUM class is returned with the same metadata, a new name and
a 'TRANS_TYPE' measure_type.

Also it was defined how the __array__ method should be handled.